add Laser callibration task
Input
The first line of the input file contains a natural number d (1 <= d <= 10), specifying the number of data sets whose descriptions are placed sequentially in the following lines of the file. The description of a single set is as follows: On its only line there are two real numbers A and B (0 < A, B < 90) denoting the angle between the mirrors and the angle between the laser beam and the mirror, respectively. These numbers are given to at most six decimal places.

Output
Each set in the input file should correspond to one line of the output file. This line should contain a single non-negative integer specifying the number of reflections of the ray from the mirrors.