Create BucketTower.cpp
In front of us there is a row of buckets. Each bucket is perfect - it is the surface of a cylinder without a top base. Each cylinder is thus fully characterized by its radius ri, and height hi.

Our task is simple. First, we take the first bucket and position it in front of us. Later we take the second one and set it (in the air for now) so that its axis of symmetry coincides with the axis of symmetry of the first cylinder, then we let it down. We do the same with the next buckets. When the job is done, there is a construction of buckets in front of us. Some of the buckets fit into others, some do not. We assume that one bucket "goes into" another if the former has a smaller radius. We also assume that all walls have zero thickness. Your task is to write a program that, for given heights and radii of successive buckets, calculates the height of a structure built this way.

The first line of the input contains a natural number d (1 <= d <= 10), specifying the number of data sets whose descriptions are placed sequentially in the following lines. The description of a single set is as follows: The first line contains a natural number n (1 <= n <= 1,000,000) specifying the number of buckets. The next line contains 2n integers: r1, h1, r2, h2, , rn, hn (1 <= ri <= 10000, 1 <= hi <= 1000). The number ri specifies the radius of the i-th bucket, while hi specifies its height.

Each set in the input should correspond to one line of the output. This line should contain a single integer specifying the height of the structure built from the buckets given in the set.