Card stacking
When playing cards, you have probably experienced the necessity of arranging the large number of cards held in your hand in a certain order. It is not uncommon to want to do this very quickly so as to not block the game.

Consider some simple rules for sorting cards. In one turn, you can remove any (exactly one) card from the fan held in your hand and put it in any other place. Your task is, for the given order of cards in the hand at the beginning, to specify in how many least steps it can be sorted in the sorted order.Translated with www.DeepL.com/Translator (free version)
Consider some simple rules for sorting cards. In one turn, you can remove any (exactly one) card from the fan held in your hand and put it in any other place. Your task is, for the given order of cards in the hand at the beginning, to specify in how many least steps it can be sorted in the sorted order.

The first line of input contains the number of test sets. Each set consists of two lines. The first contains the number of N (1 ≤ N ≤ 100000) cards held in hand. The second line contains a permutation of the set of numbers {1,2,...,N} that represents the order of the cards in hand at the very beginning.

For each set, write out in how many least steps the cards can be sorted ascendingly.