Page reading
While riding public communication it is not uncommon to read some content. Unfortunately, it often happens that they are on single pages. This causes a lot of problems. Apart from the fact that the pages may not be in the right order, some of them may be rotated by 180˚ in relation to the correct position.

We assume that all pages are printed on one side and that they are all arranged in such a way that the printed side is facing up. Such a set can be read as follows. First, if the initial page is wrongly rotated, we rotate the entire set of cards 180˚. Later, in each step, we do the following:

we read the page that is on top,
lift it above the rest
turn the remaining set (consisting of all the pages except the one you picked up) so that the top card of this set is turned correctly
without any rotation put the card you picked up on the bottom.
These actions can be performed infinitely (we neglect the fact of rereading the same cards). The natural question that arises is whether at some point all the cards will be correctly turned, and if so, after which round. The round involving the eventual turning of the entire set is numbered 1, while the next rounds are 2, 3, . . . .

Your task is to write a program that, for a given number of pages n and their initial arrangement {0/1} (0 means a page not rotated, 1 meand rotated 180˚) will determine if all the pages will ever be rotated correctly, and if so, after which round it will happen for the first time (if all of them are well arranged at the beginning, i.e., a1 = a2 = ... = an = 0, the answer is 0).

The first line of the input contains a natural number d (1 <= d <= 10), specifying the number of data sets whose descriptions are placed sequentially in the following lines of the input. The description of a single set is as follows:

The first line contains one natural number n (1 <= n <= 10000) specifying the number of cards. In the next line there is a sequence of n numbers a1, a2, . . . , an (ai belongs to {0; 1} set) specifying the initial rotation of successive pages according to the rule given in the body of the task.

Each set in the input should correspond to one line of output. This line should contain either a natural number r specifying the smallest number of rounds after which all cards will be correctly rotated (i.e., not rotated) or the word NIGDY if this situation never occurs.